refactor: extract core & shared module
Following best practices, the app should
be structured into a core module exporting mostly services and a shared
module exporting shared components,
directives and pipes.

fixes #42
