+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { UserService } from './shared';
+import { UserService } from './core';
 
 @Component({
   selector: 'app-root',
