+++ src/app/auth/no-auth-guard.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
-import { UserService } from '../shared';
+import { UserService } from '../core';
 import { map } from 'rxjs/operators/map';
 import { take } from 'rxjs/operators/take';
