+++ src/app/core/core.module.ts
@@ -1,0 +1,34 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HttpTokenInterceptor } from './interceptors/http.token.interceptor';
+
+import {
+  ApiService,
+  ArticlesService,
+  AuthGuard,
+  CommentsService,
+  JwtService,
+  ProfilesService,
+  TagsService,
+  UserService
+} from './services';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  providers: [
+    { provide: HTTP_INTERCEPTORS, useClass: HttpTokenInterceptor, multi: true },
+    ApiService,
+    ArticlesService,
+    AuthGuard,
+    CommentsService,
+    JwtService,
+    ProfilesService,
+    TagsService,
+    UserService
+  ],
+  declarations: []
+})
+export class CoreModule { }
