+++ src/app/core/index.ts
@@ -1,0 +1,4 @@
+export * from './core.module';
+export * from './services';
+export * from './models';
+export * from './interceptors';
