+++ src/app/core/interceptors/http.token.interceptor.ts
