+++ src/app/core/models/article-list-config.model.ts
