+++ src/app/core/models/article.model.ts
