+++ src/app/core/models/user.model.ts
