+++ src/app/editor/editable-article-resolver.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 
-import { Article, ArticlesService, UserService } from '../shared';
+import { Article, ArticlesService, UserService } from '../core';
 import { catchError } from 'rxjs/operators/catchError';
 import { map } from 'rxjs/operators/map';
