+++ src/app/editor/editor-routing.module.ts
@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { EditorComponent } from './editor.component';
 import { EditableArticleResolver } from './editable-article-resolver.service';
-import { AuthGuard, SharedModule } from '../shared';
+import { AuthGuard } from '../core';
+import { SharedModule } from '../shared';
 
 const routes: Routes = [
   {
