+++ src/app/editor/editor.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, FormControl } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { Article, ArticlesService } from '../shared';
+import { Article, ArticlesService } from '../core';
 
 @Component({
   selector: 'app-editor-page',
