+++ src/app/editor/editor.module.ts
@@ -3,20 +3,13 @@ import { RouterModule } from '@angular/router';
 
 import { EditorComponent } from './editor.component';
 import { EditableArticleResolver } from './editable-article-resolver.service';
-import { AuthGuard, SharedModule } from '../shared';
+import { AuthGuard } from '../core';
+import { SharedModule } from '../shared';
 import { EditorRoutingModule } from './editor-routing.module';
 
-
 @NgModule({
-  imports: [
-    SharedModule,
-    EditorRoutingModule
-  ],
-  declarations: [
-    EditorComponent
-  ],
-  providers: [
-    EditableArticleResolver
-  ]
+  imports: [SharedModule, EditorRoutingModule],
+  declarations: [EditorComponent],
+  providers: [EditableArticleResolver]
 })
 export class EditorModule {}
