+++ src/app/profile/profile-favorites.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { ArticleListConfig, Profile } from '../shared';
+import { ArticleListConfig, Profile } from '../core';
 
 @Component({
   selector: 'app-profile-favorites',
