+++ src/app/settings/settings.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 
-import { User, UserService } from '../shared';
+import { User, UserService } from '../core';
 
 @Component({
   selector: 'app-settings-page',
