+++ src/app/settings/settings.module.ts
@@ -2,7 +2,8 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 import { SettingsComponent } from './settings.component';
-import { AuthGuard, SharedModule } from '../shared';
+import { AuthGuard } from '../core';
+import { SharedModule } from '../shared';
 import { SettingsRoutingModule } from './settings-routing.module';
 
 @NgModule({
