+++ src/app/shared/index.ts
@@ -2,8 +2,5 @@ export * from './article-helpers';
 export * from './buttons';
 export * from './layout';
 export * from './list-errors.component';
-export * from './models';
-export * from './services';
 export * from './shared.module';
 export * from './show-authed.directive';
-export * from './interceptors';
