+++ src/app/shared/layout/header.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { User } from '../models';
-import { UserService } from '../services';
+import { User, UserService } from '../../core';
 
 @Component({
   selector: 'app-layout-header',
