+++ src/app/shared/list-errors.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { Errors } from './models';
+import { Errors } from '../core';
 
 @Component({
   selector: 'app-list-errors',
